Improved support for Ninja CMake generator
Using ADDITIONAL_CLEAN_FILES instead of the BYPRODUCTS
allows the artifacts from the custom target to be cleaned up
without specifying each file explicitly. These files are useful to
keep around for inspection/diagnosis so it might not be desirable to
remove them immediately after producing the required outputs. See
"CMP0058" for details on the behavior of BYPRODUCTS when using Ninja.