Updated rtos usb compatibility with lib_xud
Updated hard coded offset and removed an unneeded instruction
(register no longer has junk data alongside of PID).

Updated rtos_usb XUD_Main call.

Updated rtos_usb_endpoint_transfer_start() to include whether the
transfer is for preparation of a setup packet or not. This is to
help determine which elements in the ep_info to update and signal as
'ready for new data'.