Fixed TinyUSB and lib_xud compatibility issue
lib_xud expects an output endpoint's data buffer is large
enough to receive the endpoints data along with the CRC16;
TinyUSB does not have this same expectation and does not
offer flexibility for the control transfers on EP0. The solution
is to use an intermediate buffer and copy the received data
without the CRC16 to the TinyUSB provided buffer.

Updated rtos_usb_endpoint_transfer_start() documentation to reflect
this requirement.

Added/updated alignment attributes as another assurance.